refactoring & fix nox
